+++ src/components/DividerLine/DividerLine.stories.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import { DividerLine } from './DividerLine'
+
+export default {
+  component: DividerLine,
+  title: 'General/DividerLine',
+}
+
+export const Default = () => <DividerLine />
