+++ src/components/TextInputField/TextInputField.stories.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import TextInputField from './TextInputField'
+
+export default {
+  component: TextInputField,
+  title: 'Form Elements/TextInputField',
+}
+
+export const Default = () => <TextInputField />
